requirements.txt
openai
pandas
flask
python-dotenv
